+++ cli/src/declarations.ts
@@ -297,17 +297,6 @@ export interface CapacitorConfig {
      * @default true
      */
     resolveServiceWorkerRequests?: boolean;
-
-    /**
-     * If set to "force", margins will be adjusted for edge to edge regardless of any other settings.
-     * If set to "auto", or is missing, will check for Android 15 and the setting of [windowOptOutEdgeToEdgeEnforcement](https://developer.android.com/reference/android/R.attr#windowOptOutEdgeToEdgeEnforcement) and will adjust margins if on Android 15 and windowOptOutEdgeToEdgeEnforcement is false/missing.
-     * If set to "disable", will not adjust margins at all.
-     * In Capacitor 8, this default will be changed to 'auto'
-     *
-     * @since 7.1.0
-     * @default disable
-     */
-    adjustMarginsForEdgeToEdge?: 'auto' | 'force' | 'disable';
   };
 
   ios?: {
@@ -709,5 +698,44 @@ export interface PluginsConfig {
      * @default false
      */
     enabled?: boolean;
+  };
+
+  /**
+   * System Bars plugin configuration
+   *
+   * @since 8.0.0
+   */
+  SystemBars?: {
+    /**
+     * Disables the injection of device css insets into the web view.
+     *
+     * @default false
+     */
+    disableInsets?: boolean;
+    /**
+     * The style of the text and icons of the system bars.
+     *
+     * This option is only supported on Android.
+     *
+     * @default `DEFAULT`
+     */
+    style?: string;
+
+    /**
+     * Hide the system bars on start.
+     *
+     * @default false
+     */
+    hidden?: boolean;
+
+    /**
+     * The type of status bar animation used when showing or hiding.
+     *
+     * This option is only supported on iOS.
+     *
+     * @default 'FADE'
+     *
+     */
+    animation?: 'FADE' | 'NONE';
   };
 }